+++ src/application/wizard/service.ts
@@ -74,36 +74,9 @@ export class WizardService {
         // TODO: resolve alongside scienceBeam
         const uploadPromise = this.fileService.upload(fileContents, savedFile);
 
-        const semanticExtractionPromise = this.semanticExtractionService.extractTitle(???);
+        const semanticExtractionPromise = this.semanticExtractionService.extractTitle(fileContents, mimeType, filename);
 
-        const include = 'title';
-        const scienceBeamApiUrl = 'https://sciencebeam-texture.elifesciences.org/api/convert';
-        const scienceBeamTimeout = 20000;
-        let title = '';
-        let titleArray = 'timed-out';
-        const xmlBuffer = await axios.post(scienceBeamApiUrl, {
-            body: fileContents,
-            qs: {
-                filename,
-                include,
-            },
-            headers: { 'content-type': mimeType },
-            timeout: scienceBeamTimeout,
-        });
-
-        const parseString = promisify(xml2js.parseString);
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        const xmlData: any = await parseString(xmlBuffer.toString());
-
-        if (xmlData.article) {
-            const firstArticle = xmlData.article.front[0];
-            const articleMeta = firstArticle['article-meta'];
-            const firstMeta = articleMeta[0];
-            const titleGroup = firstMeta['title-group'];
-            const firstTitleGroup = titleGroup[0];
-            titleArray = firstTitleGroup['article-title'];
-            title = titleArray[0];
-        }
+        const results = Promise.all([uploadPromise, semanticExtractionPromise]);
 
         return savedFile;
     }
